Fixed #25 web capability at pub.dev
