--- example/lib/main.dart
@@ -1,1 +1,0 @@
-
