+++ lib/extended_image_library.dart
@@ -5,6 +5,7 @@ export 'package:http_client_helper/http_client_helper.dart';
 export 'src/extended_asset_bundle_image_provider.dart';
 export 'src/extended_file_image_provider.dart';
 export 'src/extended_image_provider.dart';
+export 'src/extended_image_utils.dart';
 export 'src/extended_memory_image_provider.dart';
 export 'src/extended_network_image_provider.dart';
 export 'src/extended_network_image_utils.dart';
